Add AM_LDFLAGS to prog_LDFLAGS for programs where it's missing
These programs override the default linker flags in order to link
with -lGLU, but as a side effect they omit linking with other
essential things like GTK+ itself.  Though GtkGlExt has a dependency
on GTK+, not all linkers are willing to search transitive
dependencies;  a program that calls GTK+ functions needs to link with
GTK+ explicitly.