Removed references to GdkDrawable.
GdkDrawable has been removed from GDK 3. With this patch GdkGLWindow
does not inherit from it any more and all references have been
removed. GdkGLWindow now directly inherits from GObject.

The API of gtkglext replicates version 1.4 of the GLX specification
in many ways, but changes the semantics for GL windows to be fully
drawable. The removal of GdkDrawable removes this drawability and
leaves the GdkGLWindow's basic functionality in place, which is the
management of ancillary buffers for OpenGL.

Gtkglext's own interface GdkGLDrawable is left mostly intact by this
patch. Possible future support for Pbuffers could implement it to
ease application development.

http://bugzilla.gnome.org/show_bug.cgi?id=642249