Added new interface to make context current.
The new interface gdk_gl_context_make_current is build around the GL
context that is to be made current. The old interfaces were build
around the drawable, but are not sufficient for setting different
different drawables for drawing and reading. Additionally, the function
gdk_gl_context_release_current has been added to release the current
GL context and any related resources.