Removed Quartz target.
The coming changes for removing the extension interface and porting
gtkglext to GTK+ 3 will bring significant changes to the internal
and external interfaces of the library. I do not have the resources
(i.e., MacOS) to adapt the Quartz backend during the transistion.

To keep the code base clean, this commit removes the Quartz backend
for now. The code might be restored of rewritten once the transition
has been done.